+++ backend/src/main/java/com/preproject/backend/domain/tag/repository/TagRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TagRepository extends JpaRepository<TagEntity, Integer> {
+public interface TagRepository extends JpaRepository<TagEntity, Long> {
     Optional<TagEntity> findByName(String name);
 
     Page<TagEntity> findByNameIncluded(String name, Pageable pageable);
